requirements.txt
Django==2
django-jinja==2.4.1
django-recaptcha2==1.0.0
django-sortedm2m==1.3.3
django-tinymce==2.4.0
django-rest-framework==0.1.0
djangorestframework-jwt==1.11.0
django-cors-headers==2.2.0
Pillow==3.2.0
pytz==2016.4
sorl-thumbnail==12.3
unidecode==0.4.19
